Move code out of `MockApi` and into utilities
